CRUD working correctly

+++ src/main/java/co/edu/uniquindio/bd1/proyectobd1/controllers/ModelFactoryController.java
@@ -28,11 +28,25 @@ public class ModelFactoryController {
     private UserServiceImp userService;
     private UserTypeServiceImp userTypeService;
 
-    private String usuarioSesion="";
+    private String usuarioSesion = "";
 
     //TODO Realizar estos tres metodos y devuelvan listas de DTO con los datos necesarios
     public List<EmployeeRegisterDTO> obtenerEmpleados() {
-        return new ArrayList<EmployeeRegisterDTO>();
+        List<EmployeeRegisterDTO> empleados = new ArrayList<EmployeeRegisterDTO>();
+        List<Employee> empleadosObtenidos = employeeService.findAll();
+        for (Employee empleado : empleadosObtenidos) {
+            empleados.add(new EmployeeRegisterDTO(
+                    empleado.getUser().getLogin(),
+                    empleado.getUser().getPassword(),
+                    empleado.getCode(),
+                    empleado.getName(),
+                    empleado.getEmail(),
+                    empleado.getEmployeePosition().getName(),
+                    empleado.getBranch().getMunicipality().getName(),
+                    empleado.getBranch().getName()
+            ));
+        }
+        return empleados;
     }
 
     public List<Object> obtenerPagosEmpleado() {
@@ -48,8 +62,8 @@ public class ModelFactoryController {
     }
 
     public List<String> obtenerMunicipios() {
-        List<Municipality> municipios=municipalityService.findAllMunicipalities();
-        List<String> nombresMunicipios=new ArrayList<String>();
+        List<Municipality> municipios = municipalityService.findAllMunicipalities();
+        List<String> nombresMunicipios = new ArrayList<String>();
         for (Municipality municipio : municipios) {
             nombresMunicipios.add(municipio.getName());
         }
@@ -57,11 +71,14 @@ public class ModelFactoryController {
     }
 
     public List<String> obtenerSucursales(String nombreMunicipio) {
-        Municipality municipio=municipalityService.findByName(nombreMunicipio).get();
-        List<Branch> sucursales=branchService.findByMunipality(municipio.getCode());
-        List<String> nombresSucursales=new ArrayList<String>();
-        for (Branch branch : sucursales) {
-            nombresSucursales.add(branch.getName());
+        List<String> nombresSucursales = new ArrayList<String>();
+        if (!nombreMunicipio.equals("")) {
+            Municipality municipio = municipalityService.findByName(nombreMunicipio).get();
+            List<Branch> sucursales = branchService.findByMunipality(municipio.getCode());
+
+            for (Branch branch : sucursales) {
+                nombresSucursales.add(branch.getName());
+            }
         }
         return nombresSucursales;
     }
@@ -73,20 +90,23 @@ public class ModelFactoryController {
         if (employeeService.findByCode(empleado.code()).isPresent()) {
             throw new Exception("Ya existe un empleado con ese codigo");
         }
-        EmployeePosition cargoEmpleado=employeePositionService.findByName(empleado.position()).get();
-        UserType tipoUsuario=userTypeService.findByEmployeePositionName(cargoEmpleado.getName()).get();
-        Branch sucursal=branchService.finByName(empleado.branch()).get();
-        User usuario=new User(empleado.login(),empleado.password(),LocalDate.now(),tipoUsuario);
+        if (employeeService.findByEmail(empleado.email()).isPresent()) {
+            throw new Exception("Ya existe un empleado con ese correo");
+        }
+        EmployeePosition cargoEmpleado = employeePositionService.findByName(empleado.position()).get();
+        UserType tipoUsuario = userTypeService.findByEmployeePositionName(cargoEmpleado.getName()).get();
+        Branch sucursal = branchService.finByName(empleado.branch()).get();
+        User usuario = new User(empleado.login(), empleado.password(), LocalDate.now(), tipoUsuario);
         userService.save(usuario);
-        Employee empleadoRegistrar=new Employee(empleado.name(),empleado.email(),
-                false,usuario,sucursal,cargoEmpleado);
+        Employee empleadoRegistrar = new Employee(empleado.name(), empleado.email(),
+                false, usuario, sucursal, cargoEmpleado);
         empleadoRegistrar.setCode(empleado.code());
         employeeService.save(empleadoRegistrar);
     }
 
     public List<String> obtenerCargosEmpleados() {
-        List<EmployeePosition> cargosEmpleados=employeePositionService.findAllEmployeePositions();
-        List<String> cargosEmpleadosNombres=new ArrayList<String>();
+        List<EmployeePosition> cargosEmpleados = employeePositionService.findAllEmployeePositions();
+        List<String> cargosEmpleadosNombres = new ArrayList<String>();
         for (EmployeePosition cargo : cargosEmpleados) {
             cargosEmpleadosNombres.add(cargo.getName());
         }
@@ -100,27 +120,37 @@ public class ModelFactoryController {
         if (!Objects.equals(empleado.oldCode(), empleado.code()) && employeeService.findByCode(empleado.code()).isPresent()) {
             throw new Exception("Ya existe un empleado con ese codigo");
         }
-        EmployeePosition cargoEmpleado=employeePositionService.findByName(empleado.position()).get();
-        UserType tipoUsuario=userTypeService.findByEmployeePositionName(cargoEmpleado.getName()).get();
-        Branch sucursal=branchService.finByName(empleado.branch()).get();
-        User usuario=userService.findByLogin(empleado.oldLogin()).get();
+        if (!empleado.oldEmail().equals(empleado.email()) && employeeService.findByEmail(empleado.email()).isPresent()) {
+            throw new Exception("Ya existe un empleado con ese correo");
+        }
+        EmployeePosition cargoEmpleado = employeePositionService.findByName(empleado.position()).get();
+        UserType tipoUsuario = userTypeService.findByEmployeePositionName(cargoEmpleado.getName()).get();
+        Branch sucursal = branchService.finByName(empleado.branch()).get();
+        User usuario = userService.findByLogin(empleado.oldLogin()).get();
+        Employee empleadoRegistrar = employeeService.findByCode(empleado.oldCode()).get();
+        employeeService.deleteEmployee(empleadoRegistrar);
+        System.out.println("0"+empleadoRegistrar.toString());
+        userService.delete(usuario);
         usuario.setLogin(empleado.login());
         usuario.setPassword(empleado.password());
         usuario.setUserType(tipoUsuario);
+        System.out.println("1"+usuario.toString());
         userService.save(usuario);
-        Employee empleadoRegistrar=employeeService.findByCode(empleado.oldCode()).get();
-        employeeService.deleteEmployee(empleadoRegistrar);
+        System.out.println("2"+usuario.toString());
+        System.out.println("3"+empleadoRegistrar.toString());
         empleadoRegistrar.setCode(empleado.code());
         empleadoRegistrar.setName(empleado.name());
         empleadoRegistrar.setEmail(empleado.email());
         empleadoRegistrar.setBranch(sucursal);
         empleadoRegistrar.setEmployeePosition(cargoEmpleado);
+        empleadoRegistrar.setUser(usuario);
+        System.out.println("4"+empleadoRegistrar.toString());
         employeeService.save(empleadoRegistrar);
     }
 
     public void borrarUsuario(EmployeeDeleteDTO empleado) {
-        User usuario=userService.findByLogin(empleado.login()).get();
-        Employee empleadoRegistrar=employeeService.findByCode(empleado.code()).get();
+        User usuario = userService.findByLogin(empleado.login()).get();
+        Employee empleadoRegistrar = employeeService.findByCode(empleado.code()).get();
         employeeService.deleteEmployee(empleadoRegistrar);
         userService.delete(usuario);
     }
@@ -136,6 +166,7 @@ public class ModelFactoryController {
 
     /**
      * Método estático para obtener la única instancia de ModelFactoryController.
+     *
      * @return La instancia única de ModelFactoryController.
      */
     public static ModelFactoryController getInstance() {
@@ -146,16 +177,6 @@ public class ModelFactoryController {
 
     }
 
-    public boolean estaBaseDatosIncompleta() {
-        boolean incompleta=false;
-        List<Employee> l =employeeService.findAll();
-        System.out.println(l.size()+" (ModelFactory)");
-        if (employeeService.findAll().isEmpty()) {
-            incompleta=true;
-        }
-        return incompleta;
-    }
-
     public void quemarDatos() {
         quemarDatosMunicipios();
         quemarSucursales();
@@ -164,25 +185,25 @@ public class ModelFactoryController {
     }
 
     public void quemarDatosMunicipios() {
-        Municipality municipio1= new Municipality("Montenegro");
-        municipio1.setCode((long)1);
-        Municipality municipio2= new Municipality("Armenia");
-        municipio2.setCode((long)2);
+        Municipality municipio1 = new Municipality("Montenegro");
+        municipio1.setCode((long) 1);
+        Municipality municipio2 = new Municipality("Armenia");
+        municipio2.setCode((long) 2);
         municipalityService.save(municipio1);
         municipalityService.save(municipio2);
     }
 
     public void quemarSucursales() {
-        Municipality municipio1=municipalityService.findByCode((long)1).get();
-        Municipality municipio2=municipalityService.findByCode((long)2).get();
-        Branch sucursal1=new Branch("Central",municipio1);
-        sucursal1.setCodeBranch((long)1);
-        Branch sucursal2=new Branch("La Perla",municipio1);
-        sucursal2.setCodeBranch((long)2);
-        Branch sucursal3=new Branch("El Sol",municipio2);
-        sucursal3.setCodeBranch((long)3);
-        Branch sucursal4=new Branch("Vista Alegre",municipio2);
-        sucursal4.setCodeBranch((long)4);
+        Municipality municipio1 = municipalityService.findByCode((long) 1).get();
+        Municipality municipio2 = municipalityService.findByCode((long) 2).get();
+        Branch sucursal1 = new Branch("Central", municipio1);
+        sucursal1.setCodeBranch((long) 1);
+        Branch sucursal2 = new Branch("La Perla", municipio1);
+        sucursal2.setCodeBranch((long) 2);
+        Branch sucursal3 = new Branch("El Sol", municipio2);
+        sucursal3.setCodeBranch((long) 3);
+        Branch sucursal4 = new Branch("Vista Alegre", municipio2);
+        sucursal4.setCodeBranch((long) 4);
         branchService.save(sucursal1);
         branchService.save(sucursal2);
         branchService.save(sucursal3);
@@ -190,16 +211,16 @@ public class ModelFactoryController {
     }
 
     public void quemarCargoEmpleado() {
-        EmployeePosition cargoEmpleado1=new EmployeePosition("Administrador",0,15000000);
-        cargoEmpleado1.setCode((long)1);
-        EmployeePosition cargoEmpleado2=new EmployeePosition("Tesoreria",0,12000000);
-        cargoEmpleado2.setCode((long)2);
-        EmployeePosition cargoEmpleado3=new EmployeePosition("Operario",0,10000000);
-        cargoEmpleado3.setCode((long)3);
-        EmployeePosition cargoEmpleado4=new EmployeePosition("Ejecutivo",0,20000000);
-        cargoEmpleado4.setCode((long)4);
-        EmployeePosition cargoEmpleado5=new EmployeePosition("Otros",0,12000000);
-        cargoEmpleado5.setCode((long)5);
+        EmployeePosition cargoEmpleado1 = new EmployeePosition("Administrador", 0, 15000000);
+        cargoEmpleado1.setCode((long) 1);
+        EmployeePosition cargoEmpleado2 = new EmployeePosition("Tesoreria", 0, 12000000);
+        cargoEmpleado2.setCode((long) 2);
+        EmployeePosition cargoEmpleado3 = new EmployeePosition("Operario", 0, 10000000);
+        cargoEmpleado3.setCode((long) 3);
+        EmployeePosition cargoEmpleado4 = new EmployeePosition("Ejecutivo", 0, 20000000);
+        cargoEmpleado4.setCode((long) 4);
+        EmployeePosition cargoEmpleado5 = new EmployeePosition("Otros", 0, 12000000);
+        cargoEmpleado5.setCode((long) 5);
         employeePositionService.save(cargoEmpleado1);
         employeePositionService.save(cargoEmpleado2);
         employeePositionService.save(cargoEmpleado3);
@@ -208,12 +229,12 @@ public class ModelFactoryController {
     }
 
     public void quemarTipoUsuario() {
-        UserType tipoUsuario1=new UserType("Administrador");
-        tipoUsuario1.setCode((long)1);
-        UserType tipoUsuario2=new UserType("Param\u00E9trico");
-        tipoUsuario2.setCode((long)2);
-        UserType tipoUsuario3=new UserType("Espor\u00E1dicos");
-        tipoUsuario3.setCode((long)3);
+        UserType tipoUsuario1 = new UserType("Administrador");
+        tipoUsuario1.setCode((long) 1);
+        UserType tipoUsuario2 = new UserType("Param\u00E9trico");
+        tipoUsuario2.setCode((long) 2);
+        UserType tipoUsuario3 = new UserType("Espor\u00E1dicos");
+        tipoUsuario3.setCode((long) 3);
         userTypeService.save(tipoUsuario1);
         userTypeService.save(tipoUsuario2);
         userTypeService.save(tipoUsuario3);
